move load bal files to simplify structure
